+++ tlm_adjoint/block_system.py
@@ -7,6 +7,8 @@ from abc import ABC, abstractmethod
 from collections.abc import Mapping, MutableMapping
 from collections import deque
 from collections.abc import Sequence
+from functools import wraps
+import logging
 try:
     import mpi4py.MPI as MPI
 except ImportError:
@@ -26,7 +28,9 @@ __all__ = \
         "BlockNullspace",
 
         "Matrix",
-        "BlockMatrix"
+        "BlockMatrix",
+
+        "System"
     ]
 
 
@@ -604,3 +608,225 @@ class Preconditioner(PETScInterface):
                 self._y, self._x)
 
         self._post_mult(y)
+
+
+class System:
+    """A linear system
+
+    .. math::
+
+        A u = b.
+
+    :arg arg_spaces: Defines the space for `u`.
+    :arg action_spaces: Defines the space for `b`.
+    :arg blocks: One of
+
+        - A :class:`tlm_adjoint.block_system.Matrix` or :class:`ufl.Form`
+          defining :math:`A`.
+        - A :class:`Mapping` with items `((i, j), block)` where the matrix
+          associated with the block in the `i` th and `j` th column is defined
+          by `block`. Each `block` is a
+          :class:`tlm_adjoint.block_system.Matrix` or :class:`ufl.Form`, or
+          `None` to indicate a zero block.
+
+    :arg nullspaces: A :class:`.Nullspace` or a :class:`Sequence` of
+        :class:`.Nullspace` objects defining the nullspace and left nullspace
+        of :math:`A`. `None` indicates a :class:`.NoneNullspace`.
+    :arg comm: Communicator.
+    """
+
+    def __init__(self, arg_spaces, action_spaces, blocks, *,
+                 nullspaces=None, comm=None):
+        if isinstance(blocks, BlockMatrix):
+            matrix = blocks
+        else:
+            matrix = BlockMatrix(arg_spaces, action_spaces, blocks)
+        arg_spaces = matrix.arg_space
+        action_spaces = matrix.action_space
+
+        nullspace = BlockNullspace(nullspaces)
+        if isinstance(nullspace, BlockNullspace):
+            if len(nullspace) != len(arg_spaces):
+                raise ValueError("Invalid space")
+            if len(nullspace) != len(action_spaces):
+                raise ValueError("Invalid space")
+
+        if comm is None:
+            comm = arg_spaces.comm
+
+        self._comm = comm
+        self._arg_space = arg_spaces
+        self._action_space = action_spaces
+        self._matrix = matrix
+        self._nullspace = nullspace
+
+    def solve(self, u, b, *,
+              solver_parameters=None, pc_fn=None,
+              pre_callback=None, post_callback=None,
+              correct_initial_guess=True, correct_solution=True):
+        """Solve the linear system.
+
+        :arg u: Defines the solution :math:`u`.
+        :arg b: Defines the right-hand-side :math:`b`.
+        :arg solver_parameters: A :class:`Mapping` defining outer Krylov solver
+            parameters. Parameters (a number of which are based on FEniCS
+            solver parameters) are:
+
+            - `'linear_solver'`: The Krylov solver type, default `'fgmres'`.
+            - `'pc_side'`: Overrides the PETSc default preconditioning side.
+            - `'relative_tolerance'`: Relative tolerance. Required.
+            - `'absolute_tolerance'`: Absolute tolerance. Required.
+            - `'divergence_limit'`: Overrides the default divergence limit.
+            - `'maximum_iterations'`: Maximum number of iterations. Default
+              1000.
+            - `'norm_type'`: Overrides the default convergence norm definition.
+            - `'nonzero_initial_guess'`: Whether to use a non-zero initial
+              guess, defined by the input `u`. Default `True`.
+            - `'gmres_restart'`: Overrides the default GMRES restart parameter.
+
+        :arg pc_fn: Defines the application of a preconditioner. A callable
+
+            .. code-block:: python
+
+                def pc_fn(u, b):
+
+            The preconditioner is applied to `b`, and the result stored in `u`.
+            Defaults to an identity.
+        :arg pre_callback: A callable accepting a single
+            :class:`petsc4py.PETSc.KSP` argument. Used for detailed manual
+            configuration. Called after all other configuration options are
+            set, but before the :meth:`petsc4py.PETSc.KSP.setUp` method is
+            called.
+        :arg post_callback: A callable accepting a single
+            :class:`petsc4py.PETSc.KSP` argument. Called after the
+            :meth:`petsc4py.PETSc.KSP.solve` method has been called.
+        :arg correct_initial_guess: Whether to apply a nullspace correction to
+            the initial guess.
+        :arg correct_solution: Whether to apply a nullspace correction to
+            the solution.
+        :returns: The number of Krylov iterations.
+        """
+
+        if solver_parameters is None:
+            solver_parameters = {}
+
+        if isinstance(u, Sequence):
+            u = tuple(u)
+        else:
+            u = (u,)
+
+            if pc_fn is not None:
+                pc_fn_u = pc_fn
+
+                @wraps(pc_fn_u)
+                def pc_fn(u, b):
+                    u, = tuple(iter_sub(u))
+                    return pc_fn_u(u, b)
+        u = self._arg_space.tuple_sub(u)
+
+        if isinstance(b, Sequence):
+            b = tuple(b)
+        else:
+            b = (b,)
+
+            if pc_fn is not None:
+                pc_fn_b = pc_fn
+
+                @wraps(pc_fn_b)
+                def pc_fn(u, b):
+                    b, = tuple(iter_sub(b))
+                    return pc_fn_b(u, b)
+        b = self._action_space.tuple_sub(b)
+
+        if tuple(u_i.function_space() for u_i in iter_sub(u)) \
+                != self._arg_space.flattened_space:
+            raise ValueError("Invalid space")
+        for b_i, space in zip_sub(b, self._action_space.split_space):
+            if b_i is not None and b_i.function_space() != space:
+                raise ValueError("Invalid space")
+
+        b_c = self._action_space.new_split()
+        for b_c_i, b_i in zip_sub(b_c, b):
+            if b_i is not None:
+                var_assign(b_c_i, b_i)
+
+        A = SystemMatrix(self._matrix, self._nullspace)
+
+        mat_A = PETSc.Mat().createPython(
+            ((self._action_space.local_size, self._action_space.global_size),
+             (self._arg_space.local_size, self._arg_space.global_size)), A,
+            comm=self._comm)
+        mat_A.setUp()
+
+        if pc_fn is not None:
+            A_pc = Preconditioner(self._action_space, self._arg_space,
+                                  pc_fn, self._nullspace)
+            pc = PETSc.PC().createPython(
+                A_pc, comm=self._comm)
+            pc.setOperators(mat_A)
+            pc.setUp()
+
+        ksp_solver = PETSc.KSP().create(comm=self._comm)
+        ksp_solver.setType(solver_parameters.get("linear_solver", "fgmres"))
+        if pc_fn is not None:
+            ksp_solver.setPC(pc)
+        if "pc_side" in solver_parameters:
+            ksp_solver.setPCSide(solver_parameters["pc_side"])
+        ksp_solver.setOperators(mat_A)
+        ksp_solver.setTolerances(
+            rtol=solver_parameters["relative_tolerance"],
+            atol=solver_parameters["absolute_tolerance"],
+            divtol=solver_parameters.get("divergence_limit", None),
+            max_it=solver_parameters.get("maximum_iterations", 1000))
+        ksp_solver.setInitialGuessNonzero(
+            solver_parameters.get("nonzero_initial_guess", True))
+        ksp_solver.setNormType(
+            solver_parameters.get(
+                "norm_type", PETSc.KSP.NormType.DEFAULT))
+        if "gmres_restart" in solver_parameters:
+            ksp_solver.setGMRESRestart(solver_parameters["gmres_restart"])
+
+        logger = logging.getLogger("tlm_adjoint.System")
+
+        def monitor(ksp_solver, it, r_norm):
+            logger.debug(f"KSP: "
+                         f"iteration {it:d}, "
+                         f"residual norm {r_norm:.16e}")
+
+        ksp_solver.setMonitor(monitor)
+
+        if correct_initial_guess:
+            self._nullspace.correct_soln(u)
+        self._nullspace.correct_rhs(b_c)
+
+        u_petsc = mat_A.createVecRight()
+        self._arg_space.to_petsc(u_petsc, u)
+        b_petsc = mat_A.createVecLeft()
+        self._action_space.to_petsc(b_petsc, b_c)
+        del b_c
+
+        if pre_callback is not None:
+            pre_callback(ksp_solver)
+        ksp_solver.setUp()
+        ksp_solver.solve(b_petsc, u_petsc)
+        if post_callback is not None:
+            post_callback(ksp_solver)
+        del b_petsc
+
+        self._arg_space.from_petsc(u_petsc, u)
+        del u_petsc
+
+        if correct_solution:
+            # Not needed if the linear problem were to be solved exactly
+            self._nullspace.correct_soln(u)
+
+        if ksp_solver.getConvergedReason() <= 0:
+            raise RuntimeError("Convergence failure")
+        ksp_its = ksp_solver.getIterationNumber()
+
+        ksp_solver.destroy()
+        mat_A.destroy()
+        if pc_fn is not None:
+            pc.destroy()
+
+        return ksp_its
