+++ tlm_adjoint/firedrake/__init__.py
@@ -9,8 +9,6 @@ from .backend import backend, backend_RealType, backend_ScalarType  # noqa: E402
 from .backend_interface import *    # noqa: E402,F401
 from .assembly import *             # noqa: E402,F401
 from .assignment import *           # noqa: E402,F401
-from .block_system import \
-    (ConstantNullspace, DirichletBCNullspace, NoneNullspace, UnityNullspace)  # noqa: E402,E501,F401
 from .caches import *               # noqa: E402,F401
 from .expr import *                 # noqa: E402,F401
 from .hessian_system import *       # noqa: E402,F401
@@ -22,6 +20,6 @@ from .variables import *            # noqa: E402,F401
 
 from .backend_patches import *      # noqa: E402,F401
 
-del backend, backend_interface, assembly, assignment, block_system, caches, \
-    expr, hessian_system, interpolation, parameters, projection, solve, \
-    variables, backend_patches  # noqa: F821
+del backend, backend_interface, assembly, assignment, caches, expr, \
+    hessian_system, interpolation, parameters, projection, solve, variables, \
+    backend_patches  # noqa: F821
